nyc_schools/backend/requirements.txt
flask
python-dotenv
requests
simplejson
pytest
psycopg2
sodapy

